Extract axis from chart
