+++ frontend/scripts/react-components/chart/chart.component.jsx
@@ -1,29 +1,22 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 
-import maxBy from 'lodash/maxBy';
-import max from 'lodash/max';
-
 import {
   Line,
   Bar,
   Cell,
   Area,
   Pie,
-  XAxis,
-  YAxis,
   CartesianGrid,
   Tooltip,
   Legend,
   ResponsiveContainer,
   ComposedChart,
-  PieChart,
-  Label
+  PieChart
 } from 'recharts';
 
-import ChartTick from 'react-components/chart/tick/tick.component';
-import CategoryTick from 'react-components/chart/tick/category-tick.component';
-
+import CustomXAxis from 'react-components/chart/x-axis.component';
+import CustomYAxis from 'react-components/chart/y-axis.component';
 import 'react-components/chart/chart-styles.scss';
 
 class Chart extends PureComponent {
@@ -40,85 +33,6 @@ class Chart extends PureComponent {
     handleMouseMove: null,
     handleMouseLeave: null
   };
-
-  findMaxValue = (data, axisKeys) => {
-    const maxValues = [];
-    Object.keys(axisKeys).forEach(key => {
-      Object.keys(axisKeys[key]).forEach(subKey => {
-        maxValues.push(maxBy(data, subKey)[subKey]);
-      });
-    });
-    return max(maxValues);
-  };
-
-  renderXAxis() {
-    const { config, data } = this.props;
-    const { xKey, xKeys, xAxis, layout, unit } = config;
-    if (!xAxis) return null;
-    if (layout === 'vertical') {
-      const maxXValue = this.findMaxValue(data, xKeys);
-      return (
-        <XAxis
-          axisLine={false}
-          tickLine={false}
-          tickMargin={10}
-          tick={
-            <ChartTick
-              dataMax={maxXValue}
-              unitFormat={value => value}
-              unit={unit}
-              backgroundColor="transparent"
-              fill="white"
-            />
-          }
-          {...xAxis}
-        >
-          <Label position="insideBottom" offset={-15} className="x-axis-label">
-            {unit}
-          </Label>
-        </XAxis>
-      );
-    }
-
-    return <XAxis dataKey={xKey || ''} tick={{ fontSize: 12 }} {...xAxis} />;
-  }
-
-  renderYAxis() {
-    const { config, data } = this.props;
-    const { yAxis, unit, unitFormat, yKeys, yKey } = config;
-    if (yKey)
-      return (
-        <YAxis
-          axisLine={false}
-          tickLine={false}
-          tickMargin={15}
-          dataKey={yKey || ''}
-          tick={<CategoryTick />}
-          {...yAxis}
-        />
-      );
-    const maxYValue = this.findMaxValue(data, yKeys);
-    return (
-      yAxis && (
-        <YAxis
-          axisLine={false}
-          tickSize={-50}
-          mirror
-          tickMargin={0}
-          tickLine={false}
-          tick={
-            <ChartTick
-              dataMax={maxYValue}
-              unit={unit || ''}
-              unitFormat={unitFormat || (value => value)}
-              fill="#555555"
-            />
-          }
-          {...yAxis}
-        />
-      )
-    );
-  }
 
   render() {
     const { className, data, config, handleMouseMove, handleMouseLeave } = this.props;
@@ -202,8 +116,8 @@ class Chart extends PureComponent {
             {cartesianGrid && (
               <CartesianGrid strokeDasharray="4 4" stroke="#d6d6d9" {...cartesianGrid} />
             )}
-            {this.renderXAxis()}
-            {this.renderYAxis()}
+            {CustomXAxis({ config, data })}
+            {CustomYAxis({ config, data })}
             {areas &&
               Object.keys(areas).map(key => (
                 <Area key={key} dataKey={key} dot={false} {...areas[key]} />
